pyproject.toml
[project]
name = "training-tracker-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.23.2",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
]
